% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_domain_plot.R
\name{normalized_domain_plot}
\alias{normalized_domain_plot}
\title{Create a Normalized Domain Plot}
\usage{
normalized_domain_plot(featured_col, train_data, test_data, treshold_value)
}
\arguments{
\item{featured_col}{The name of the featured column.}

\item{train_data}{A data frame containing the training data.}

\item{test_data}{A data frame containing the test data.}

\item{treshold_value}{The threshold value for the domain plot.}
}
\value{
A domain plot visualizing the distances of imputed values.
}
\description{
This function generates a domain plot for a normalized model based on PCA
distances of the provided data.
}
\examples{
set.seed(123)
library(dplyr)
# Specifying featured column
featured_col = "cd_2022"
train_data = viral |>
  dplyr::select("cd_2022", "vl_2022")
test_data = sero
treshold_value = 0.99
# Call the function
}
