% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sero}
\alias{sero}
\title{Seropositive Data for Applicability Domain Testing}
\format{
A tibble (data frame) with 53 rows and 2 columns.
}
\usage{
data(sero)
}
\description{
This dataset is designed for testing the applicability domain of methods
related to HIV research. It provides a tibble with 53 rows and 2 columns
containing numeric measurements of CD4 lymphocyte counts (cd_2022) and viral
load (vl_2022) for seropositive individuals in 2022. These measurements are
vital indicators of HIV disease status. This dataset is ideal for evaluating
the performance and suitability of various HIV-predictive models and as an
aid in developing diagnostic tools within a seropositive context.
}
\note{
To explore more rows of this dataset, you can use the \code{print(n = ...)} function.
}
\examples{
data(sero)
sero
}
\author{
Juan Pablo Acuña González \href{mailto:acua6307@gmail.com}{acua6307@gmail.com}
}
\keyword{datasets}
