% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_domain_plot.R
\name{simple_domain_plot}
\alias{simple_domain_plot}
\title{Create a Simple Domain Plot}
\usage{
simple_domain_plot(featured_col, train_data, test_data, treshold_value)
}
\arguments{
\item{featured_col}{Name of the featured column in the training data.}

\item{train_data}{The training data used to fit the model.}

\item{test_data}{The testing domain data used to calculate PCA distances.}

\item{treshold_value}{The threshold for domain applicability scoring.}
}
\value{
A simple damain plot
}
\description{
This function generates a domain plot for a simple model based on PCA
distances of the provided data.
}
\examples{
set.seed(123)
library(dplyr)
# Specifying featured column
featured_col = "cd_2022"
train_data = viral |>
  dplyr::select("cd_2022", "vl_2022")
test_data = sero
treshold_value = 0.99
# Call the function
}
