% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-data.R
\docType{data}
\name{train}
\alias{train}
\title{Training Data for Applicability Domain Analysis}
\format{
A tibble (data frame) with 26 rows and 6 columns.
}
\usage{
data(train)
}
\description{
This dataset contains training data for viral load models applicability
domain analysis. It includes CD4 and viral load measurements for different years.
}
\note{
To explore more rows of this dataset, you can use the \code{print(n = ...)} function.
}
\examples{
data(train)
train
}
\author{
Juan Pablo Acuña González \href{mailto:22253567@uagro.mx}{22253567@uagro.mx}
}
\keyword{datasets}
