% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viral}
\alias{viral}
\title{Predictive Modeling Data for Viral Load and CD4 Lymphocyte Counts}
\format{
A tibble (data frame) with 35 rows and 6 columns.
}
\usage{
data(viral)
}
\description{
This dataset serves as input for predictive modeling tasks related to HIV
research. It contains numeric measurements of CD4 lymphocyte counts (cd) and
viral load (vl) at three different time points: 2019, 2021, and 2022. These
measurements are crucial indicators of HIV disease progression.
}
\note{
To explore more rows of this dataset, you can use the \code{print(n = ...)} function.
}
\examples{
data(viral)
viral
}
\author{
Juan Pablo Acuña González \href{mailto:acua6307@gmail.com}{acua6307@gmail.com}
}
\keyword{datasets}
